move processing utility function under processing/utils
