+++ tests/unit/processing/test_color_conversion_utils.py
@@ -2,7 +2,7 @@ import unittest
 
 import anndata
 
-from backend.common.utils.color_conversion_utils import (
+from backend.layers.processing.utils.color_conversion_utils import (
     ColorFormatException,
     convert_anndata_category_colors_to_cxg_category_colors,
     convert_color_to_hex_format,
